pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cs454-team14-proj"
version = "0.1.0"
description = "PE-focused CNN accelerator mapping optimization using Timeloop, ONNX, and NSGA-II"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "onnx>=1.12.0",
    "pymoo>=0.6.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pe_mapper"]
